update: data structure for ParallelWLC

+++ include/webcachesim/caches/parallel_wlc.h
@@ -277,7 +277,8 @@ public:
 
     ParallelInCacheLRUQueue in_cache_lru_queue;
     //TODO: negative queue should have a better abstraction, at least hide the round-up
-    sparse_hash_map<uint32_t, uint64_t> negative_candidate_queue;
+    //sparse_hash_map<uint64_t, uint64_t> has smaller memory overhead than <uint32_t, uint64_t>
+    sparse_hash_map<uint64_t, uint64_t> negative_candidate_queue;
     ParallelWLCTrainingData *training_data;
     ParallelWLCTrainingData *background_training_data;
     std::mutex training_data_mutex;
